Don't push if no dead domains removed

+++ workflow_scripts/dead.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 
 domains_file="domains"
-email="91372088+jarelllama@users.noreply.github.com"
-name="jarelllama"
+github_email="91372088+jarelllama@users.noreply.github.com"
+github_name="jarelllama"
 
 grep -vE '^(#|$)' "$domains_file" > tmp1.txt
 
 touch tmp_dead.txt
+
+echo -e "\nDead domains:"
 
 cat tmp1.txt | xargs -I{} -P8 bash -c "
   if dig @1.1.1.1 {} | grep -Fq 'NXDOMAIN'; then
@@ -15,16 +17,19 @@ cat tmp1.txt | xargs -I{} -P8 bash -c "
   fi
 "
 
-comm -23 tmp1.txt <(sort tmp_dead.txt) > "$domains_file"
+if ! [[ -s tmp_dead.txt ]]; then
+    echo -e "\nNo dead domains found.\n"
+    exit 0
+fi
 
 echo -e "\nTotal domains removed: $(wc -l < tmp_dead.txt)\n"
 
-update_header
+comm -23 tmp1.txt <(sort tmp_dead.txt) > "$domains_file"
 
 rm tmp*.txt
 
-git config user.email "$email"
-git config user.name "$name"
+git config user.email "$github_email"
+git config user.name "$github_name"
 
 git add "$domains_file"
 git commit -m "Remove dead domains"
